@@ -1,5 +1,5 @@
-describe("Server.ts tests", () => {
-    test("Math test", () => {
-      expect(2 + 2).toBe(4);
-    });
+describe('Server.ts tests', () => {
+  test('Math test', () => {
+    expect(2 + 2).toBe(4);
   });
+});
