@@ -1,12 +1,13 @@
-import express from "express";
-import { addProduct, getAllProducts, getSingleProduct} from '../controllers/productsControllers';
+import express from 'express';
+import {
+  addProduct,
+  getAllProducts,
+  getSingleProduct,
+} from '../controllers/productsControllers';
 
 const router = express.Router();
 
-
-
-router.route("/").get(getAllProducts).post(addProduct);
-router.route("/:id").get(getSingleProduct);
-
+router.route('/').get(getAllProducts).post(addProduct);
+router.route('/:id').get(getSingleProduct);
 
 export default router;
