@@ -1,9 +1,9 @@
-import { Router, Request, Response } from "express";
+import { Router, Request, Response } from 'express';
 
 const router = Router();
 
-router.get("/", (req: Request, res: Response): void => {
-  let users = ["Goon", "Tsuki", "Joe"];
+router.get('/', (req: Request, res: Response): void => {
+  const users = ['Goon', 'Tsuki', 'Joe'];
   res.status(200).send(users);
 });
 
